Add swipeable actions to recycler actions
I was testing some functionality where I was using the scroll state
changed listener of the recyclerview to do a certain action. The scroll*
functions do not trigger a state change since these immediately scroll
to a certain position. The swiping functions allows you to emulate the
action of dragging with your finger across the screen, however these
were not available on the recycler view.